create horizontal carousel for posters

+++ feature/discover/src/main/java/com/buntupana/tmdb/feature/discover/domain/model/MovieItem.kt
@@ -10,7 +10,7 @@ data class MovieItem(
     val originalLanguage: String,
     val genreIds: List<Int>,
     val popularity: Double,
-    val voteAverage: Double,
+    val voteAverage: Int,
     val voteCount: Int,
     val releaseDate: String,
     val video: Boolean,
